Removed duplicate scripts
